requirements.txt
# =============================================================================
# FDTM Platform - Python Dependencies
# =============================================================================

# Core Django
Django>=5.0,<6.0
python-dotenv>=1.0.0
django-environ>=0.11.2
Pillow>=10.0.0

# Database
dj-database-url>=2.1.0

# Forms and Admin
django-crispy-forms>=2.1
crispy-tailwind>=1.0.0

# Rich Text Editor
django-ckeditor-5>=0.2.10

# Storage - Backblaze B2
django-storages>=1.14.0
boto3>=1.34.0

# Payments
stripe>=7.0.0

# Translation
deepl>=1.16.0

# Image optimization
django-imagekit>=5.0.0

# HTMX support (optional but recommended)
django-htmx>=1.17.0

# Development
django-debug-toolbar>=4.2.0

# Production
gunicorn>=21.0.0
whitenoise>=6.6.0
